pyproject.toml
[tool.poetry]
name = "seqsampling"
version = "0.1.0"
description = "study of sampling"
authors = [
  "Sergey Troshin <serj.troshin@proton.me>",
  "Irina Saparina <i.saparina@sms.ed.ac.uk>",
]
readme = "README.md"
package-mode = true

# Use src layout: code in src/seqsampling
packages = [
    { include = "seqsampling", from = "src" }
]

[tool.poetry.dependencies]
python       = "3.10.15"
pkginfo      = ">=1.12"
# Serving / eval
tqdm = "4.66.1"
pytest = ">=7.4.0,<9.0.0"
hydra-core = "^1.3.2"
pyyaml = "^6.0.2"
openai       = "^2.8.1"

# NEW: needed for the sequential sampling infra
httpx        = "^0.27.0"   # HTTP client for VLLMClient
uvicorn      = "^0.32.0"   # to run vLLM server wrapper
pydantic     = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^5.0.0"
